decided upon a tabbed layout for ease of future features, nonfunctional WIP UI
